refactor: rename worker.worker to worker.impl
