+++ config.example.ini
@@ -27,7 +27,7 @@
 
 [worker]
 # Fully qualified name of the worker class
-#type = questionpy_server.worker.worker.subprocess.SubprocessWorker
+#type = questionpy_server.worker.impl.subprocess.SubprocessWorker
 
 # Maximum amount of running workers
 #max_workers = 8
