+++ questionpy_server/web/_routes/_packages.py
@@ -16,7 +16,7 @@ from questionpy_server.web.app import QPyServer
 from questionpy_server.worker.runtime.package_location import ZipPackageLocation
 
 if TYPE_CHECKING:
-    from questionpy_server.worker.worker import Worker
+    from questionpy_server.worker import Worker
 
 package_routes = web.RouteTableDef()
