+++ questionpy_server/worker/impl/thread.py
@@ -13,11 +13,11 @@ from collections.abc import Sequence
 from questionpy_common.environment import WorkerResourceLimits
 from questionpy_server.worker.connection import ServerToWorkerConnection
 from questionpy_server.worker.exception import WorkerNotRunningError
+from questionpy_server.worker.impl._base import BaseWorker
 from questionpy_server.worker.runtime.connection import WorkerToServerConnection
 from questionpy_server.worker.runtime.manager import WorkerManager
 from questionpy_server.worker.runtime.package_location import PackageLocation
 from questionpy_server.worker.runtime.streams import AsyncReadAdapter, DuplexPipe
-from questionpy_server.worker.worker.base import BaseWorker
 
 log = logging.getLogger(__name__)
