+++ tests/questionpy_server/worker/impl/test_subprocess.py
@@ -9,7 +9,7 @@ import pytest
 
 from questionpy_common.constants import MiB
 from questionpy_server import WorkerPool
-from questionpy_server.worker.worker.subprocess import SubprocessWorker
+from questionpy_server.worker.impl.subprocess import SubprocessWorker
 from tests.conftest import PACKAGE
