+++ tests/questionpy_server/worker/impl/test_thread.py
@@ -11,9 +11,9 @@ import pytest
 from questionpy_common.constants import MiB
 from questionpy_server import WorkerPool
 from questionpy_server.worker.exception import WorkerStartError
+from questionpy_server.worker.impl.thread import ThreadWorker
 from questionpy_server.worker.runtime.manager import WorkerManager
 from questionpy_server.worker.runtime.messages import WorkerUnknownError
-from questionpy_server.worker.worker.thread import ThreadWorker
 from tests.conftest import PACKAGE
